Add documentation for each service
I would like to highlight what each service does, briefly but a bit more
detailed compared with the single line we have in the docs homepage.

I created a page for each service (excluded pbnj for now), I hope it
will be also a way to start getting more of it written for each service.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>