templates: Move task info closer to before example Template
I think it makes sense to introduce the relation ship of Template:task
and task:action before showing an example. Having this information
before the example should set the reader up to make better sense of the
yaml they are looking at since they have more context.

The rest of the information should be easy enough to figure out for
anyone with docker/docker-compose experience.

Signed-off-by: Manuel Mendez <mmendez534@gmail.com>