Update naming of Hegel to Tootles
This commit updates the name of the metadata service from Hegel to
Tootles.

There is one reference to Hegel left, in the example output in the
Getting Started page. This will be updated later when I had a chance
to replay the Getting Started page and fetch fresh output.

I updated all references to the old Hegel repo to the tootles/ directory
in the tinkerbell/tinkerbell repo for now. In the next step, I will port
the Hegel repo's README there.

Signed-off-by: Michael Meier <mmeier1986@gmail.com>