Update links throughout the docs
The main part of this commit is changing all links to point to the
consolidated tinkerbell/tinkerbell repository where components have
been moved there.

Where feasible, I introduced a "latestTinkerbellVersion" variable
pointing to the GitHub tree of the 0.18.3 release, and then pointed
all links to files into that tree.

I changed the links to the individual repos on the homepage to point
to their docs pages instead, as that seems more "representative" than
a directory with source code in the consolidated repo.

Signed-off-by: Michael Meier <mmeier1986@gmail.com>