Introduce shortcode for GH repo tree refs
The new repo_tree shortcode can be used to reference a path in the
repository given under params.tinkerbellRepo at the tag given under
params.tinkerbellVersion.

And example for referencing the README.md at v0.18.3 would look like
this:
  {{< repo_tree "README.md" >}}

This shortcode is now being used in all places where files in the Tinkerbell
repository are referenced. This means that when an update happens, only the
tinkerbellVersion needs to be updated in hugo.toml to update all references
to files in the repo.

Signed-off-by: Michael Meier <mmeier1986@gmail.com>